Simplify database migrations and fix CORS duplicate entries
