+++ backend/alembic/versions/23372f1dd158_create_all_tables.py
@@ -1,0 +1,842 @@
+"""create_all_tables\
+
+Revision ID: 23372f1dd158
+Revises: 
+Create Date: 2025-04-16 02:11:54.380042
+
+"""
+
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "23372f1dd158"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "team",
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("slug", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("avatar_url", sa.String(length=1024), nullable=True),
+        sa.Column("team_size", sa.Integer(), nullable=False),
+        sa.Column("is_personal", sa.Boolean(), nullable=False),
+        sa.Column(
+            "team_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("created_by_user_id", sa.String(length=255), nullable=False),
+        sa.Column("created_by_email", sa.String(length=255), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_team_id"), "team", ["id"], unique=False)
+    op.create_index(op.f("ix_team_slug"), "team", ["slug"], unique=False)
+    op.create_index("ix_team_slug_unique", "team", ["slug"], unique=True)
+    op.create_table(
+        "integration",
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column(
+            "service_type",
+            sa.Enum("SLACK", "GITHUB", "NOTION", "DISCORD", name="integrationtype"),
+            nullable=False,
+        ),
+        sa.Column(
+            "status",
+            sa.Enum(
+                "ACTIVE",
+                "DISCONNECTED",
+                "EXPIRED",
+                "REVOKED",
+                "ERROR",
+                name="integrationstatus",
+            ),
+            nullable=False,
+        ),
+        sa.Column(
+            "integration_metadata",
+            postgresql.JSONB(astext_type=sa.Text()),
+            nullable=True,
+        ),
+        sa.Column("last_used_at", sa.DateTime(), nullable=True),
+        sa.Column("owner_team_id", sa.UUID(), nullable=False),
+        sa.Column("created_by_user_id", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["owner_team_id"],
+            ["team.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_integration_id"), "integration", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_integration_owner_team_id"),
+        "integration",
+        ["owner_team_id"],
+        unique=False,
+    )
+    op.create_table(
+        "slackworkspace",
+        sa.Column("slack_id", sa.String(length=255), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("domain", sa.String(length=255), nullable=True),
+        sa.Column("icon_url", sa.String(length=1024), nullable=True),
+        sa.Column("team_size", sa.Integer(), nullable=True),
+        sa.Column(
+            "workspace_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("is_connected", sa.Boolean(), nullable=False),
+        sa.Column("connection_status", sa.String(length=50), nullable=False),
+        sa.Column("last_connected_at", sa.DateTime(), nullable=False),
+        sa.Column("last_sync_at", sa.DateTime(), nullable=True),
+        sa.Column("access_token", sa.String(length=1024), nullable=True),
+        sa.Column("refresh_token", sa.String(length=1024), nullable=True),
+        sa.Column("token_expires_at", sa.DateTime(), nullable=True),
+        sa.Column("team_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["team_id"],
+            ["team.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_slackworkspace_id"), "slackworkspace", ["id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_slackworkspace_slack_id"), "slackworkspace", ["slack_id"], unique=True
+    )
+    op.create_index(
+        op.f("ix_slackworkspace_team_id"), "slackworkspace", ["team_id"], unique=False
+    )
+    op.create_table(
+        "teammember",
+        sa.Column("user_id", sa.String(length=255), nullable=False),
+        sa.Column("email", sa.String(length=255), nullable=True),
+        sa.Column("display_name", sa.String(length=255), nullable=True),
+        sa.Column(
+            "role",
+            sa.Enum("OWNER", "ADMIN", "MEMBER", "VIEWER", name="teammemberrole"),
+            nullable=False,
+        ),
+        sa.Column("invitation_status", sa.String(length=50), nullable=False),
+        sa.Column("invitation_token", sa.String(length=255), nullable=True),
+        sa.Column("invitation_expires_at", sa.DateTime(), nullable=True),
+        sa.Column("last_active_at", sa.DateTime(), nullable=True),
+        sa.Column("team_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["team_id"],
+            ["team.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_teammember_id"), "teammember", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_teammember_team_id"), "teammember", ["team_id"], unique=False
+    )
+    op.create_index(
+        "ix_teammember_team_id_user_id",
+        "teammember",
+        ["team_id", "user_id"],
+        unique=True,
+    )
+    op.create_index(
+        op.f("ix_teammember_user_id"), "teammember", ["user_id"], unique=False
+    )
+    op.create_table(
+        "integration_share",
+        sa.Column(
+            "share_level",
+            sa.Enum("FULL_ACCESS", "LIMITED_ACCESS", "READ_ONLY", name="sharelevel"),
+            nullable=False,
+        ),
+        sa.Column("status", sa.String(length=50), nullable=False),
+        sa.Column("revoked_at", sa.DateTime(), nullable=True),
+        sa.Column("integration_id", sa.UUID(), nullable=False),
+        sa.Column("team_id", sa.UUID(), nullable=False),
+        sa.Column("shared_by_user_id", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["integration_id"],
+            ["integration.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["team_id"],
+            ["team.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_integration_share_id"), "integration_share", ["id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_integration_share_integration_id"),
+        "integration_share",
+        ["integration_id"],
+        unique=False,
+    )
+    op.create_index(
+        "ix_integration_share_integration_id_team_id",
+        "integration_share",
+        ["integration_id", "team_id"],
+        unique=True,
+    )
+    op.create_index(
+        op.f("ix_integration_share_team_id"),
+        "integration_share",
+        ["team_id"],
+        unique=False,
+    )
+    op.create_table(
+        "integrationcredential",
+        sa.Column(
+            "credential_type",
+            sa.Enum(
+                "OAUTH_TOKEN",
+                "PERSONAL_TOKEN",
+                "API_KEY",
+                "APP_TOKEN",
+                name="credentialtype",
+            ),
+            nullable=False,
+        ),
+        sa.Column("encrypted_value", sa.String(length=2048), nullable=False),
+        sa.Column("expires_at", sa.DateTime(), nullable=True),
+        sa.Column("refresh_token", sa.String(length=2048), nullable=True),
+        sa.Column("scopes", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+        sa.Column("integration_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["integration_id"],
+            ["integration.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_integrationcredential_id"),
+        "integrationcredential",
+        ["id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_integrationcredential_integration_id"),
+        "integrationcredential",
+        ["integration_id"],
+        unique=False,
+    )
+    op.create_table(
+        "integrationevent",
+        sa.Column(
+            "event_type",
+            sa.Enum(
+                "CREATED",
+                "SHARED",
+                "UNSHARED",
+                "UPDATED",
+                "DISCONNECTED",
+                "ACCESS_CHANGED",
+                "ERROR",
+                name="eventtype",
+            ),
+            nullable=False,
+        ),
+        sa.Column("details", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+        sa.Column("integration_id", sa.UUID(), nullable=False),
+        sa.Column("actor_user_id", sa.String(length=255), nullable=False),
+        sa.Column("affected_team_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["affected_team_id"],
+            ["team.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["integration_id"],
+            ["integration.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_integrationevent_affected_team_id"),
+        "integrationevent",
+        ["affected_team_id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_integrationevent_id"), "integrationevent", ["id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_integrationevent_integration_id"),
+        "integrationevent",
+        ["integration_id"],
+        unique=False,
+    )
+    op.create_table(
+        "serviceresource",
+        sa.Column(
+            "resource_type",
+            sa.Enum(
+                "SLACK_CHANNEL",
+                "SLACK_USER",
+                "SLACK_EMOJI",
+                "GITHUB_REPOSITORY",
+                "GITHUB_ISSUE",
+                "GITHUB_PR",
+                "GITHUB_WEBHOOK",
+                "NOTION_PAGE",
+                "NOTION_DATABASE",
+                "NOTION_BLOCK",
+                "DISCORD_GUILD",
+                "DISCORD_CHANNEL",
+                name="resourcetype",
+            ),
+            nullable=False,
+        ),
+        sa.Column("external_id", sa.String(length=255), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column(
+            "resource_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("last_synced_at", sa.DateTime(), nullable=True),
+        sa.Column("integration_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["integration_id"],
+            ["integration.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_serviceresource_id"), "serviceresource", ["id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_serviceresource_integration_id"),
+        "serviceresource",
+        ["integration_id"],
+        unique=False,
+    )
+    op.create_index(
+        "ix_serviceresource_integration_id_resource_type_external_id",
+        "serviceresource",
+        ["integration_id", "resource_type", "external_id"],
+        unique=True,
+    )
+    op.create_table(
+        "slackanalysis",
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("start_date", sa.DateTime(), nullable=False),
+        sa.Column("end_date", sa.DateTime(), nullable=False),
+        sa.Column("parameters", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+        sa.Column("llm_model", sa.String(length=255), nullable=True),
+        sa.Column("analysis_type", sa.String(length=50), nullable=False),
+        sa.Column("is_scheduled", sa.Boolean(), nullable=False),
+        sa.Column("schedule_frequency", sa.String(length=50), nullable=True),
+        sa.Column("next_run_at", sa.DateTime(), nullable=True),
+        sa.Column("status", sa.String(length=50), nullable=False),
+        sa.Column("progress", sa.Float(), nullable=False),
+        sa.Column("error_message", sa.Text(), nullable=True),
+        sa.Column(
+            "result_summary", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("completion_time", sa.DateTime(), nullable=True),
+        sa.Column("workspace_id", sa.UUID(), nullable=False),
+        sa.Column("created_by_user_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["workspace_id"],
+            ["slackworkspace.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_slackanalysis_id"), "slackanalysis", ["id"], unique=False)
+    op.create_table(
+        "slackchannel",
+        sa.Column("slack_id", sa.String(length=255), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("type", sa.String(length=50), nullable=False),
+        sa.Column("purpose", sa.String(length=1024), nullable=True),
+        sa.Column("topic", sa.String(length=1024), nullable=True),
+        sa.Column("member_count", sa.Integer(), nullable=True),
+        sa.Column("is_archived", sa.Boolean(), nullable=False),
+        sa.Column("created_at_ts", sa.String(length=50), nullable=True),
+        sa.Column("is_bot_member", sa.Boolean(), nullable=False),
+        sa.Column("bot_joined_at", sa.DateTime(), nullable=True),
+        sa.Column("is_selected_for_analysis", sa.Boolean(), nullable=False),
+        sa.Column("is_supported", sa.Boolean(), nullable=False),
+        sa.Column("last_sync_at", sa.DateTime(), nullable=True),
+        sa.Column("oldest_synced_ts", sa.String(length=50), nullable=True),
+        sa.Column("latest_synced_ts", sa.String(length=50), nullable=True),
+        sa.Column("workspace_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["workspace_id"],
+            ["slackworkspace.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_slackchannel_id"), "slackchannel", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_slackchannel_slack_id"), "slackchannel", ["slack_id"], unique=False
+    )
+    op.create_index(
+        "ix_slackchannel_workspace_id_slack_id",
+        "slackchannel",
+        ["workspace_id", "slack_id"],
+        unique=True,
+    )
+    op.create_table(
+        "slackuser",
+        sa.Column("slack_id", sa.String(length=255), nullable=False),
+        sa.Column("name", sa.String(length=255), nullable=False),
+        sa.Column("display_name", sa.String(length=255), nullable=True),
+        sa.Column("real_name", sa.String(length=255), nullable=True),
+        sa.Column("email", sa.String(length=255), nullable=True),
+        sa.Column("title", sa.String(length=255), nullable=True),
+        sa.Column("phone", sa.String(length=50), nullable=True),
+        sa.Column("timezone", sa.String(length=100), nullable=True),
+        sa.Column("timezone_offset", sa.Integer(), nullable=True),
+        sa.Column("profile_image_url", sa.String(length=1024), nullable=True),
+        sa.Column("is_bot", sa.Boolean(), nullable=False),
+        sa.Column("is_admin", sa.Boolean(), nullable=False),
+        sa.Column("is_deleted", sa.Boolean(), nullable=False),
+        sa.Column(
+            "profile_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("workspace_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["workspace_id"],
+            ["slackworkspace.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_slackuser_id"), "slackuser", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_slackuser_slack_id"), "slackuser", ["slack_id"], unique=False
+    )
+    op.create_index(
+        "ix_slackuser_workspace_id_slack_id",
+        "slackuser",
+        ["workspace_id", "slack_id"],
+        unique=True,
+    )
+    op.create_table(
+        "analysis_channels",
+        sa.Column("analysis_id", sa.UUID(), nullable=False),
+        sa.Column("channel_id", sa.UUID(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["analysis_id"],
+            ["slackanalysis.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["channel_id"],
+            ["slackchannel.id"],
+        ),
+        sa.PrimaryKeyConstraint("analysis_id", "channel_id"),
+    )
+    op.create_table(
+        "resourceaccess",
+        sa.Column(
+            "access_level",
+            sa.Enum("READ", "WRITE", "ADMIN", name="accesslevel"),
+            nullable=False,
+        ),
+        sa.Column("resource_id", sa.UUID(), nullable=False),
+        sa.Column("team_id", sa.UUID(), nullable=False),
+        sa.Column("granted_by_user_id", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["resource_id"],
+            ["serviceresource.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["team_id"],
+            ["team.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_resourceaccess_id"), "resourceaccess", ["id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_resourceaccess_resource_id"),
+        "resourceaccess",
+        ["resource_id"],
+        unique=False,
+    )
+    op.create_index(
+        "ix_resourceaccess_resource_id_team_id",
+        "resourceaccess",
+        ["resource_id", "team_id"],
+        unique=True,
+    )
+    op.create_index(
+        op.f("ix_resourceaccess_team_id"), "resourceaccess", ["team_id"], unique=False
+    )
+    op.create_table(
+        "slackchannelanalysis",
+        sa.Column("analysis_id", sa.UUID(), nullable=False),
+        sa.Column("channel_id", sa.UUID(), nullable=False),
+        sa.Column("start_date", sa.DateTime(), nullable=False),
+        sa.Column("end_date", sa.DateTime(), nullable=False),
+        sa.Column("message_count", sa.Integer(), nullable=False),
+        sa.Column("participant_count", sa.Integer(), nullable=False),
+        sa.Column("thread_count", sa.Integer(), nullable=False),
+        sa.Column("reaction_count", sa.Integer(), nullable=False),
+        sa.Column("channel_summary", sa.Text(), nullable=True),
+        sa.Column("topic_analysis", sa.Text(), nullable=True),
+        sa.Column("contributor_insights", sa.Text(), nullable=True),
+        sa.Column("key_highlights", sa.Text(), nullable=True),
+        sa.Column("model_used", sa.String(length=255), nullable=True),
+        sa.Column("generated_at", sa.DateTime(), nullable=False),
+        sa.Column(
+            "raw_response", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("status", sa.String(length=50), nullable=False),
+        sa.Column("error_message", sa.Text(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["analysis_id"],
+            ["slackanalysis.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["channel_id"],
+            ["slackchannel.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_slackchannelanalysis_analysis_id"),
+        "slackchannelanalysis",
+        ["analysis_id"],
+        unique=False,
+    )
+    op.create_index(
+        "ix_slackchannelanalysis_analysis_id_channel_id",
+        "slackchannelanalysis",
+        ["analysis_id", "channel_id"],
+        unique=True,
+    )
+    op.create_index(
+        op.f("ix_slackchannelanalysis_channel_id"),
+        "slackchannelanalysis",
+        ["channel_id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_slackchannelanalysis_generated_at"),
+        "slackchannelanalysis",
+        ["generated_at"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_slackchannelanalysis_id"), "slackchannelanalysis", ["id"], unique=False
+    )
+    op.create_table(
+        "slackcontribution",
+        sa.Column("problem_solving_score", sa.Float(), nullable=True),
+        sa.Column("knowledge_sharing_score", sa.Float(), nullable=True),
+        sa.Column("team_coordination_score", sa.Float(), nullable=True),
+        sa.Column("engagement_score", sa.Float(), nullable=True),
+        sa.Column("total_score", sa.Float(), nullable=True),
+        sa.Column("message_count", sa.Integer(), nullable=False),
+        sa.Column("thread_reply_count", sa.Integer(), nullable=False),
+        sa.Column("reaction_given_count", sa.Integer(), nullable=False),
+        sa.Column("reaction_received_count", sa.Integer(), nullable=False),
+        sa.Column(
+            "notable_contributions",
+            postgresql.JSONB(astext_type=sa.Text()),
+            nullable=True,
+        ),
+        sa.Column("insights", sa.Text(), nullable=True),
+        sa.Column(
+            "insights_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("analysis_id", sa.UUID(), nullable=False),
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("channel_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["analysis_id"],
+            ["slackanalysis.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["channel_id"],
+            ["slackchannel.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["slackuser.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_slackcontribution_analysis_id_user_id_channel_id",
+        "slackcontribution",
+        ["analysis_id", "user_id", "channel_id"],
+        unique=True,
+    )
+    op.create_index(
+        op.f("ix_slackcontribution_id"), "slackcontribution", ["id"], unique=False
+    )
+    op.create_table(
+        "slackmessage",
+        sa.Column("slack_id", sa.String(length=255), nullable=False),
+        sa.Column("slack_ts", sa.String(length=50), nullable=False),
+        sa.Column("text", sa.Text(), nullable=True),
+        sa.Column("processed_text", sa.Text(), nullable=True),
+        sa.Column("message_type", sa.String(length=50), nullable=False),
+        sa.Column("subtype", sa.String(length=50), nullable=True),
+        sa.Column("is_edited", sa.Boolean(), nullable=False),
+        sa.Column("edited_ts", sa.String(length=50), nullable=True),
+        sa.Column("has_attachments", sa.Boolean(), nullable=False),
+        sa.Column(
+            "attachments", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("files", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+        sa.Column("thread_ts", sa.String(length=50), nullable=True),
+        sa.Column("is_thread_parent", sa.Boolean(), nullable=False),
+        sa.Column("is_thread_reply", sa.Boolean(), nullable=False),
+        sa.Column("reply_count", sa.Integer(), nullable=False),
+        sa.Column("reply_users_count", sa.Integer(), nullable=False),
+        sa.Column("reaction_count", sa.Integer(), nullable=False),
+        sa.Column("message_datetime", sa.DateTime(), nullable=False),
+        sa.Column("is_analyzed", sa.Boolean(), nullable=False),
+        sa.Column("message_category", sa.String(length=100), nullable=True),
+        sa.Column("sentiment_score", sa.Float(), nullable=True),
+        sa.Column(
+            "analysis_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
+        ),
+        sa.Column("channel_id", sa.UUID(), nullable=False),
+        sa.Column("user_id", sa.UUID(), nullable=True),
+        sa.Column("parent_id", sa.UUID(), nullable=True),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["channel_id"],
+            ["slackchannel.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["parent_id"],
+            ["slackmessage.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["slackuser.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        "ix_slackmessage_channel_id_slack_ts",
+        "slackmessage",
+        ["channel_id", "slack_ts"],
+        unique=False,
+    )
+    op.create_index(op.f("ix_slackmessage_id"), "slackmessage", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_slackmessage_message_datetime"),
+        "slackmessage",
+        ["message_datetime"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_slackmessage_slack_id"), "slackmessage", ["slack_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_slackmessage_slack_ts"), "slackmessage", ["slack_ts"], unique=False
+    )
+    op.create_index(
+        op.f("ix_slackmessage_thread_ts"), "slackmessage", ["thread_ts"], unique=False
+    )
+    op.create_index(
+        "ix_slackmessage_user_id_slack_ts",
+        "slackmessage",
+        ["user_id", "slack_ts"],
+        unique=False,
+    )
+    op.create_table(
+        "slackreaction",
+        sa.Column("emoji_name", sa.String(length=255), nullable=False),
+        sa.Column("emoji_code", sa.String(length=255), nullable=True),
+        sa.Column("reaction_ts", sa.String(length=50), nullable=True),
+        sa.Column("message_id", sa.UUID(), nullable=False),
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["message_id"],
+            ["slackmessage.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["slackuser.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_slackreaction_id"), "slackreaction", ["id"], unique=False)
+    op.create_index(
+        "ix_slackreaction_message_id_user_id_emoji_name",
+        "slackreaction",
+        ["message_id", "user_id", "emoji_name"],
+        unique=True,
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(
+        "ix_slackreaction_message_id_user_id_emoji_name", table_name="slackreaction"
+    )
+    op.drop_index(op.f("ix_slackreaction_id"), table_name="slackreaction")
+    op.drop_table("slackreaction")
+    op.drop_index("ix_slackmessage_user_id_slack_ts", table_name="slackmessage")
+    op.drop_index(op.f("ix_slackmessage_thread_ts"), table_name="slackmessage")
+    op.drop_index(op.f("ix_slackmessage_slack_ts"), table_name="slackmessage")
+    op.drop_index(op.f("ix_slackmessage_slack_id"), table_name="slackmessage")
+    op.drop_index(op.f("ix_slackmessage_message_datetime"), table_name="slackmessage")
+    op.drop_index(op.f("ix_slackmessage_id"), table_name="slackmessage")
+    op.drop_index("ix_slackmessage_channel_id_slack_ts", table_name="slackmessage")
+    op.drop_table("slackmessage")
+    op.drop_index(op.f("ix_slackcontribution_id"), table_name="slackcontribution")
+    op.drop_index(
+        "ix_slackcontribution_analysis_id_user_id_channel_id",
+        table_name="slackcontribution",
+    )
+    op.drop_table("slackcontribution")
+    op.drop_index(op.f("ix_slackchannelanalysis_id"), table_name="slackchannelanalysis")
+    op.drop_index(
+        op.f("ix_slackchannelanalysis_generated_at"), table_name="slackchannelanalysis"
+    )
+    op.drop_index(
+        op.f("ix_slackchannelanalysis_channel_id"), table_name="slackchannelanalysis"
+    )
+    op.drop_index(
+        "ix_slackchannelanalysis_analysis_id_channel_id",
+        table_name="slackchannelanalysis",
+    )
+    op.drop_index(
+        op.f("ix_slackchannelanalysis_analysis_id"), table_name="slackchannelanalysis"
+    )
+    op.drop_table("slackchannelanalysis")
+    op.drop_index(op.f("ix_resourceaccess_team_id"), table_name="resourceaccess")
+    op.drop_index("ix_resourceaccess_resource_id_team_id", table_name="resourceaccess")
+    op.drop_index(op.f("ix_resourceaccess_resource_id"), table_name="resourceaccess")
+    op.drop_index(op.f("ix_resourceaccess_id"), table_name="resourceaccess")
+    op.drop_table("resourceaccess")
+    op.drop_table("analysis_channels")
+    op.drop_index("ix_slackuser_workspace_id_slack_id", table_name="slackuser")
+    op.drop_index(op.f("ix_slackuser_slack_id"), table_name="slackuser")
+    op.drop_index(op.f("ix_slackuser_id"), table_name="slackuser")
+    op.drop_table("slackuser")
+    op.drop_index("ix_slackchannel_workspace_id_slack_id", table_name="slackchannel")
+    op.drop_index(op.f("ix_slackchannel_slack_id"), table_name="slackchannel")
+    op.drop_index(op.f("ix_slackchannel_id"), table_name="slackchannel")
+    op.drop_table("slackchannel")
+    op.drop_index(op.f("ix_slackanalysis_id"), table_name="slackanalysis")
+    op.drop_table("slackanalysis")
+    op.drop_index(
+        "ix_serviceresource_integration_id_resource_type_external_id",
+        table_name="serviceresource",
+    )
+    op.drop_index(
+        op.f("ix_serviceresource_integration_id"), table_name="serviceresource"
+    )
+    op.drop_index(op.f("ix_serviceresource_id"), table_name="serviceresource")
+    op.drop_table("serviceresource")
+    op.drop_index(
+        op.f("ix_integrationevent_integration_id"), table_name="integrationevent"
+    )
+    op.drop_index(op.f("ix_integrationevent_id"), table_name="integrationevent")
+    op.drop_index(
+        op.f("ix_integrationevent_affected_team_id"), table_name="integrationevent"
+    )
+    op.drop_table("integrationevent")
+    op.drop_index(
+        op.f("ix_integrationcredential_integration_id"),
+        table_name="integrationcredential",
+    )
+    op.drop_index(
+        op.f("ix_integrationcredential_id"), table_name="integrationcredential"
+    )
+    op.drop_table("integrationcredential")
+    op.drop_index(op.f("ix_integration_share_team_id"), table_name="integration_share")
+    op.drop_index(
+        "ix_integration_share_integration_id_team_id", table_name="integration_share"
+    )
+    op.drop_index(
+        op.f("ix_integration_share_integration_id"), table_name="integration_share"
+    )
+    op.drop_index(op.f("ix_integration_share_id"), table_name="integration_share")
+    op.drop_table("integration_share")
+    op.drop_index(op.f("ix_teammember_user_id"), table_name="teammember")
+    op.drop_index("ix_teammember_team_id_user_id", table_name="teammember")
+    op.drop_index(op.f("ix_teammember_team_id"), table_name="teammember")
+    op.drop_index(op.f("ix_teammember_id"), table_name="teammember")
+    op.drop_table("teammember")
+    op.drop_index(op.f("ix_slackworkspace_team_id"), table_name="slackworkspace")
+    op.drop_index(op.f("ix_slackworkspace_slack_id"), table_name="slackworkspace")
+    op.drop_index(op.f("ix_slackworkspace_id"), table_name="slackworkspace")
+    op.drop_table("slackworkspace")
+    op.drop_index(op.f("ix_integration_owner_team_id"), table_name="integration")
+    op.drop_index(op.f("ix_integration_id"), table_name="integration")
+    op.drop_table("integration")
+    op.drop_index("ix_team_slug_unique", table_name="team")
+    op.drop_index(op.f("ix_team_slug"), table_name="team")
+    op.drop_index(op.f("ix_team_id"), table_name="team")
+    op.drop_table("team")
+    # ### end Alembic commands ###
