--- backend/alembic/versions/consolidated_schema.py
@@ -1,1254 +1,0 @@
-"""Consolidated schema migration
-
-Revision ID: consolidated_schema
-Revises: 002_add_team_integrations_models
-Create Date: 2025-04-16 12:00:00.000000
-
-This is a consolidated migration that defines the entire database schema
-with proper idempotency checks for all tables.
-"""
-
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "consolidated_schema"
-down_revision = "002_add_team_integrations_models"
-branch_labels = None
-depends_on = None
-
-
-# Global dictionary to keep track of which enums exist
-_enums_created = {}
-
-
-def create_enum_if_not_exists(name, values):
-    """Create an enum type if it doesn't already exist"""
-    global _enums_created
-
-    conn = op.get_bind()
-
-    # Check if the enum exists
-    enum_exists_query = """
-    SELECT 1 FROM pg_type WHERE typname = :enum_name
-    """
-    result = conn.execute(sa.text(enum_exists_query), {"enum_name": name})
-    exists = result.scalar() is not None
-
-    if not exists:
-        # Create enum vals string like 'val1', 'val2', 'val3'
-        enum_vals = ", ".join(f"'{val}'" for val in values)
-
-        # Create the enum
-        conn.execute(sa.text(f"CREATE TYPE {name} AS ENUM ({enum_vals})"))
-        print(f"Created enum {name}")
-    else:
-        print(f"Enum {name} already exists")
-
-    # Store existence status for later use
-    _enums_created[name] = exists
-
-    return not exists
-
-
-def create_table_if_not_exists(table_name, table_definition_func):
-    """Wrapper function to create table if it doesn't exist"""
-    conn = op.get_bind()
-
-    # Check if the table exists
-    table_exists_query = """
-    SELECT 1 FROM information_schema.tables 
-    WHERE table_name = :table_name AND table_schema = 'public'
-    """
-    result = conn.execute(sa.text(table_exists_query), {"table_name": table_name})
-    exists = result.scalar() is not None
-
-    if not exists:
-        # Call the function that creates the table
-        table_definition_func()
-        print(f"Created table {table_name}")
-    else:
-        print(f"Table {table_name} already exists")
-
-    return not exists
-
-
-def upgrade():
-    """Upgrade database with idempotent schema creation"""
-
-    # Create Enums
-    enums = {
-        "teammemberrole": ["OWNER", "ADMIN", "MEMBER", "VIEWER"],
-        "integrationtype": ["slack", "github", "notion", "discord"],
-        "integrationstatus": ["active", "disconnected", "expired", "revoked", "error"],
-        "credentialtype": ["oauth_token", "personal_token", "api_key", "app_token"],
-        "sharelevel": ["full_access", "limited_access", "read_only"],
-        "resourcetype": [
-            "slack_channel",
-            "slack_user",
-            "slack_emoji",
-            "github_repository",
-            "github_issue",
-            "github_pr",
-            "github_webhook",
-            "notion_page",
-            "notion_database",
-            "notion_block",
-            "discord_guild",
-            "discord_channel",
-        ],
-        "accesslevel": ["read", "write", "admin"],
-        "eventtype": [
-            "created",
-            "shared",
-            "unshared",
-            "updated",
-            "disconnected",
-            "access_changed",
-            "error",
-        ],
-    }
-
-    for enum_name, enum_values in enums.items():
-        create_enum_if_not_exists(enum_name, enum_values)
-
-    # Create Team Table
-    def create_team_table():
-        op.create_table(
-            "team",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("slug", sa.String(255), nullable=False),
-            sa.Column("description", sa.Text(), nullable=True),
-            sa.Column("avatar_url", sa.String(1024), nullable=True),
-            sa.Column("team_size", sa.Integer(), nullable=False),
-            sa.Column("is_personal", sa.Boolean(), nullable=False),
-            sa.Column(
-                "team_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("created_by_user_id", sa.String(255), nullable=False),
-            sa.Column("created_by_email", sa.String(255), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(op.f("ix_team_id"), "team", ["id"], unique=False)
-        op.create_index(op.f("ix_team_slug"), "team", ["slug"], unique=False)
-        op.create_index("ix_team_slug_unique", "team", ["slug"], unique=True)
-
-    # Create TeamMember Table
-    def create_teammember_table():
-        op.create_table(
-            "teammember",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("user_id", sa.String(255), nullable=False),
-            sa.Column("email", sa.String(255), nullable=True),
-            sa.Column("display_name", sa.String(255), nullable=True),
-            sa.Column(
-                "role",
-                sa.Enum("OWNER", "ADMIN", "MEMBER", "VIEWER", name="teammemberrole"),
-                nullable=False,
-            ),
-            sa.Column("invitation_status", sa.String(50), nullable=False),
-            sa.Column("invitation_token", sa.String(255), nullable=True),
-            sa.Column("invitation_expires_at", sa.DateTime(), nullable=True),
-            sa.Column("last_active_at", sa.DateTime(), nullable=True),
-            sa.Column("team_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(op.f("ix_teammember_id"), "teammember", ["id"], unique=False)
-        op.create_index(
-            op.f("ix_teammember_user_id"), "teammember", ["user_id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_teammember_team_id"), "teammember", ["team_id"], unique=False
-        )
-        op.create_index(
-            "ix_teammember_team_id_user_id",
-            "teammember",
-            ["team_id", "user_id"],
-            unique=True,
-        )
-
-    # Create Integration Table
-    def create_integration_table():
-        op.create_table(
-            "integration",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("description", sa.Text(), nullable=True),
-            sa.Column(
-                "service_type",
-                sa.Enum("slack", "github", "notion", "discord", name="integrationtype"),
-                nullable=False,
-            ),
-            sa.Column(
-                "status",
-                sa.Enum(
-                    "active",
-                    "disconnected",
-                    "expired",
-                    "revoked",
-                    "error",
-                    name="integrationstatus",
-                ),
-                nullable=False,
-            ),
-            sa.Column(
-                "integration_metadata",
-                postgresql.JSONB(astext_type=sa.Text()),
-                nullable=True,
-            ),
-            sa.Column("last_used_at", sa.DateTime(), nullable=True),
-            sa.Column("owner_team_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("created_by_user_id", sa.String(255), nullable=False),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["owner_team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(op.f("ix_integration_id"), "integration", ["id"], unique=False)
-        op.create_index(
-            op.f("ix_integration_owner_team_id"),
-            "integration",
-            ["owner_team_id"],
-            unique=False,
-        )
-
-    # Create IntegrationCredential Table
-    def create_integration_credential_table():
-        # Check if we already handled the credential_type enum
-        global _enums_created
-        credential_type_exists = _enums_created.get("credentialtype", False)
-
-        # Prepare column spec for credential_type with or without enum creation
-        if credential_type_exists:
-            # If the enum already exists, use the existing enum without trying to create it
-            credential_type_col = sa.Column(
-                "credential_type",
-                sa.Enum(
-                    "oauth_token",
-                    "personal_token",
-                    "api_key",
-                    "app_token",
-                    name="credentialtype",
-                    create_type=False,
-                ),
-                nullable=False,
-            )
-        else:
-            # If the enum doesn't exist
-            credential_type_col = sa.Column(
-                "credential_type",
-                sa.Enum(
-                    "oauth_token",
-                    "personal_token",
-                    "api_key",
-                    "app_token",
-                    name="credentialtype",
-                ),
-                nullable=False,
-            )
-
-        op.create_table(
-            "integration_credential",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=False),
-            credential_type_col,
-            sa.Column(
-                "credential_data",
-                postgresql.JSONB(astext_type=sa.Text()),
-                nullable=False,
-            ),
-            sa.Column("expires_at", sa.DateTime(), nullable=True),
-            sa.Column("scopes", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["integration_id"], ["integration.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_integration_credential_id"),
-            "integration_credential",
-            ["id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_integration_credential_integration_id"),
-            "integration_credential",
-            ["integration_id"],
-            unique=False,
-        )
-
-    # Create IntegrationShare Table
-    def create_integration_share_table():
-        op.create_table(
-            "integration_share",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("team_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column(
-                "share_level",
-                sa.Enum(
-                    "full_access", "limited_access", "read_only", name="sharelevel"
-                ),
-                nullable=False,
-            ),
-            sa.Column(
-                "status",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'active'"),
-            ),
-            sa.Column("shared_by_user_id", sa.String(255), nullable=False),
-            sa.Column("revoked_by_user_id", sa.String(255), nullable=True),
-            sa.Column("revoked_at", sa.DateTime(), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["integration_id"], ["integration.id"]),
-            sa.ForeignKeyConstraint(["team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_integration_share_id"), "integration_share", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_integration_share_integration_id"),
-            "integration_share",
-            ["integration_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_integration_share_team_id"),
-            "integration_share",
-            ["team_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_integration_share_integration_id_team_id",
-            "integration_share",
-            ["integration_id", "team_id"],
-            unique=True,
-        )
-
-    # Create ServiceResource Table
-    def create_service_resource_table():
-        # Check if we already handled the resource_type enum
-        global _enums_created
-        resource_type_exists = _enums_created.get("resourcetype", False)
-
-        # Prepare column spec for resource_type with or without enum creation
-        if resource_type_exists:
-            # If the enum already exists, use the existing enum without trying to create it
-            resource_type_col = sa.Column(
-                "resource_type",
-                sa.Enum(
-                    "slack_channel",
-                    "slack_user",
-                    "slack_emoji",
-                    "github_repository",
-                    "github_issue",
-                    "github_pr",
-                    "github_webhook",
-                    "notion_page",
-                    "notion_database",
-                    "notion_block",
-                    "discord_guild",
-                    "discord_channel",
-                    name="resourcetype",
-                    create_type=False,
-                ),
-                nullable=False,
-            )
-        else:
-            # If the enum doesn't exist
-            resource_type_col = sa.Column(
-                "resource_type",
-                sa.Enum(
-                    "slack_channel",
-                    "slack_user",
-                    "slack_emoji",
-                    "github_repository",
-                    "github_issue",
-                    "github_pr",
-                    "github_webhook",
-                    "notion_page",
-                    "notion_database",
-                    "notion_block",
-                    "discord_guild",
-                    "discord_channel",
-                    name="resourcetype",
-                ),
-                nullable=False,
-            )
-
-        op.create_table(
-            "service_resource",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=False),
-            resource_type_col,
-            sa.Column("external_id", sa.String(255), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column(
-                "resource_metadata",
-                postgresql.JSONB(astext_type=sa.Text()),
-                nullable=True,
-            ),
-            sa.Column("last_synced_at", sa.DateTime(), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["integration_id"], ["integration.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_service_resource_id"), "service_resource", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_service_resource_integration_id"),
-            "service_resource",
-            ["integration_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_service_resource_integration_id_resource_type_external_id",
-            "service_resource",
-            ["integration_id", "resource_type", "external_id"],
-            unique=True,
-        )
-
-    # Create ResourceAccess Table
-    def create_resource_access_table():
-        op.create_table(
-            "resource_access",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("resource_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("team_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column(
-                "access_level",
-                sa.Enum("read", "write", "admin", name="accesslevel"),
-                nullable=False,
-            ),
-            sa.Column("granted_by_user_id", sa.String(255), nullable=False),
-            sa.Column("revoked_by_user_id", sa.String(255), nullable=True),
-            sa.Column("revoked_at", sa.DateTime(), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["resource_id"], ["service_resource.id"]),
-            sa.ForeignKeyConstraint(["team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_resource_access_id"), "resource_access", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_resource_access_resource_id"),
-            "resource_access",
-            ["resource_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_resource_access_team_id"),
-            "resource_access",
-            ["team_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_resource_access_resource_id_team_id",
-            "resource_access",
-            ["resource_id", "team_id"],
-            unique=True,
-        )
-
-    # Create IntegrationEvent Table
-    def create_integration_event_table():
-        # Check if we already handled the event_type enum
-        global _enums_created
-        event_type_exists = _enums_created.get("eventtype", False)
-
-        # Prepare column spec for event_type with or without enum creation
-        if event_type_exists:
-            # If the enum already exists, use the existing enum without trying to create it
-            event_type_col = sa.Column(
-                "event_type",
-                sa.Enum(
-                    "created",
-                    "shared",
-                    "unshared",
-                    "updated",
-                    "disconnected",
-                    "access_changed",
-                    "error",
-                    name="eventtype",
-                    create_type=False,
-                ),
-                nullable=False,
-            )
-        else:
-            # If the enum doesn't exist (shouldn't happen, but just in case)
-            event_type_col = sa.Column(
-                "event_type",
-                sa.Enum(
-                    "created",
-                    "shared",
-                    "unshared",
-                    "updated",
-                    "disconnected",
-                    "access_changed",
-                    "error",
-                    name="eventtype",
-                ),
-                nullable=False,
-            )
-
-        op.create_table(
-            "integration_event",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=False),
-            event_type_col,
-            sa.Column(
-                "details", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("actor_user_id", sa.String(255), nullable=False),
-            sa.Column("affected_team_id", postgresql.UUID(as_uuid=True), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["integration_id"], ["integration.id"]),
-            sa.ForeignKeyConstraint(["affected_team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_integration_event_id"), "integration_event", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_integration_event_integration_id"),
-            "integration_event",
-            ["integration_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_integration_event_affected_team_id"),
-            "integration_event",
-            ["affected_team_id"],
-            unique=False,
-        )
-
-    # Create Slack Tables
-    def create_slack_tables():
-        # SlackWorkspace Table
-        op.create_table(
-            "slackworkspace",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("slack_id", sa.String(255), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("domain", sa.String(255), nullable=True),
-            sa.Column("icon_url", sa.String(1024), nullable=True),
-            sa.Column("team_size", sa.Integer(), nullable=True),
-            sa.Column(
-                "workspace_metadata",
-                postgresql.JSONB(astext_type=sa.Text()),
-                nullable=True,
-            ),
-            sa.Column(
-                "is_connected",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.Column(
-                "connection_status",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'active'"),
-            ),
-            sa.Column("last_connected_at", sa.DateTime(), nullable=False),
-            sa.Column("last_sync_at", sa.DateTime(), nullable=True),
-            sa.Column("access_token", sa.String(1024), nullable=True),
-            sa.Column("refresh_token", sa.String(1024), nullable=True),
-            sa.Column("token_expires_at", sa.DateTime(), nullable=True),
-            sa.Column("team_id", postgresql.UUID(as_uuid=True), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["team_id"], ["team.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackworkspace_id"), "slackworkspace", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackworkspace_slack_id"),
-            "slackworkspace",
-            ["slack_id"],
-            unique=True,
-        )
-        op.create_index(
-            op.f("ix_slackworkspace_team_id"),
-            "slackworkspace",
-            ["team_id"],
-            unique=False,
-        )
-
-        # SlackChannel Table
-        op.create_table(
-            "slackchannel",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("slack_id", sa.String(255), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("type", sa.String(50), nullable=False),
-            sa.Column("purpose", sa.String(1024), nullable=True),
-            sa.Column("topic", sa.String(1024), nullable=True),
-            sa.Column("member_count", sa.Integer(), nullable=True),
-            sa.Column(
-                "is_archived",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column("created_at_ts", sa.String(50), nullable=True),
-            sa.Column(
-                "is_bot_member",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column("bot_joined_at", sa.DateTime(), nullable=True),
-            sa.Column(
-                "is_selected_for_analysis",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "is_supported",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.Column("last_sync_at", sa.DateTime(), nullable=True),
-            sa.Column("oldest_synced_ts", sa.String(50), nullable=True),
-            sa.Column("latest_synced_ts", sa.String(50), nullable=True),
-            sa.Column("workspace_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["workspace_id"], ["slackworkspace.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackchannel_id"), "slackchannel", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackchannel_slack_id"), "slackchannel", ["slack_id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackchannel_workspace_id"),
-            "slackchannel",
-            ["workspace_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_slackchannel_workspace_id_slack_id",
-            "slackchannel",
-            ["workspace_id", "slack_id"],
-            unique=True,
-        )
-
-        # SlackUser Table
-        op.create_table(
-            "slackuser",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("slack_id", sa.String(255), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("display_name", sa.String(255), nullable=True),
-            sa.Column("real_name", sa.String(255), nullable=True),
-            sa.Column("email", sa.String(255), nullable=True),
-            sa.Column("title", sa.String(255), nullable=True),
-            sa.Column("phone", sa.String(50), nullable=True),
-            sa.Column("timezone", sa.String(100), nullable=True),
-            sa.Column("timezone_offset", sa.Integer(), nullable=True),
-            sa.Column("profile_image_url", sa.String(1024), nullable=True),
-            sa.Column(
-                "is_bot", sa.Boolean(), nullable=False, server_default=sa.text("false")
-            ),
-            sa.Column(
-                "is_admin",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "is_deleted",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "profile_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("workspace_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["workspace_id"], ["slackworkspace.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(op.f("ix_slackuser_id"), "slackuser", ["id"], unique=False)
-        op.create_index(
-            op.f("ix_slackuser_slack_id"), "slackuser", ["slack_id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackuser_workspace_id"),
-            "slackuser",
-            ["workspace_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_slackuser_workspace_id_slack_id",
-            "slackuser",
-            ["workspace_id", "slack_id"],
-            unique=True,
-        )
-
-        # Create associations table for many-to-many relationship
-        op.create_table(
-            "analysis_channels",
-            sa.Column("analysis_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("channel_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.ForeignKeyConstraint(["analysis_id"], ["slackanalysis.id"]),
-            sa.ForeignKeyConstraint(["channel_id"], ["slackchannel.id"]),
-            sa.PrimaryKeyConstraint("analysis_id", "channel_id"),
-        )
-
-        # SlackAnalysis Table
-        op.create_table(
-            "slackanalysis",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("name", sa.String(255), nullable=False),
-            sa.Column("description", sa.Text(), nullable=True),
-            sa.Column("start_date", sa.DateTime(), nullable=False),
-            sa.Column("end_date", sa.DateTime(), nullable=False),
-            sa.Column(
-                "parameters", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("llm_model", sa.String(255), nullable=True),
-            sa.Column(
-                "analysis_type",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'channel_analysis'"),
-            ),
-            sa.Column(
-                "is_scheduled",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column("schedule_frequency", sa.String(50), nullable=True),
-            sa.Column("next_run_at", sa.DateTime(), nullable=True),
-            sa.Column(
-                "status",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'pending'"),
-            ),
-            sa.Column(
-                "progress", sa.Float(), nullable=False, server_default=sa.text("0.0")
-            ),
-            sa.Column("error_message", sa.Text(), nullable=True),
-            sa.Column(
-                "result_summary", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("completion_time", sa.DateTime(), nullable=True),
-            sa.Column("workspace_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column(
-                "created_by_user_id", postgresql.UUID(as_uuid=True), nullable=True
-            ),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["workspace_id"], ["slackworkspace.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackanalysis_id"), "slackanalysis", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackanalysis_workspace_id"),
-            "slackanalysis",
-            ["workspace_id"],
-            unique=False,
-        )
-
-        # SlackChannelAnalysis Table
-        op.create_table(
-            "slackchannelanalysis",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("analysis_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("channel_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("start_date", sa.DateTime(), nullable=False),
-            sa.Column("end_date", sa.DateTime(), nullable=False),
-            sa.Column(
-                "message_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "participant_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "thread_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "reaction_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column("channel_summary", sa.Text(), nullable=True),
-            sa.Column("topic_analysis", sa.Text(), nullable=True),
-            sa.Column("contributor_insights", sa.Text(), nullable=True),
-            sa.Column("key_highlights", sa.Text(), nullable=True),
-            sa.Column("model_used", sa.String(255), nullable=True),
-            sa.Column("generated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "raw_response", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column(
-                "status",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'completed'"),
-            ),
-            sa.Column("error_message", sa.Text(), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["analysis_id"], ["slackanalysis.id"]),
-            sa.ForeignKeyConstraint(["channel_id"], ["slackchannel.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackchannelanalysis_id"),
-            "slackchannelanalysis",
-            ["id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackchannelanalysis_analysis_id"),
-            "slackchannelanalysis",
-            ["analysis_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackchannelanalysis_channel_id"),
-            "slackchannelanalysis",
-            ["channel_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackchannelanalysis_generated_at"),
-            "slackchannelanalysis",
-            ["generated_at"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_slackchannelanalysis_analysis_id_channel_id",
-            "slackchannelanalysis",
-            ["analysis_id", "channel_id"],
-            unique=True,
-        )
-
-        # SlackContribution Table
-        op.create_table(
-            "slackcontribution",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("problem_solving_score", sa.Float(), nullable=True),
-            sa.Column("knowledge_sharing_score", sa.Float(), nullable=True),
-            sa.Column("team_coordination_score", sa.Float(), nullable=True),
-            sa.Column("engagement_score", sa.Float(), nullable=True),
-            sa.Column("total_score", sa.Float(), nullable=True),
-            sa.Column(
-                "message_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "thread_reply_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "reaction_given_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "reaction_received_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "notable_contributions",
-                postgresql.JSONB(astext_type=sa.Text()),
-                nullable=True,
-            ),
-            sa.Column("insights", sa.Text(), nullable=True),
-            sa.Column(
-                "insights_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("analysis_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("user_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("channel_id", postgresql.UUID(as_uuid=True), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["analysis_id"], ["slackanalysis.id"]),
-            sa.ForeignKeyConstraint(["user_id"], ["slackuser.id"]),
-            sa.ForeignKeyConstraint(["channel_id"], ["slackchannel.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackcontribution_id"), "slackcontribution", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackcontribution_analysis_id"),
-            "slackcontribution",
-            ["analysis_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackcontribution_user_id"),
-            "slackcontribution",
-            ["user_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackcontribution_channel_id"),
-            "slackcontribution",
-            ["channel_id"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_slackcontribution_analysis_id_user_id_channel_id",
-            "slackcontribution",
-            ["analysis_id", "user_id", "channel_id"],
-            unique=True,
-        )
-
-        # SlackMessage Table
-        op.create_table(
-            "slackmessage",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("slack_id", sa.String(255), nullable=False),
-            sa.Column("slack_ts", sa.String(50), nullable=False),
-            sa.Column("text", sa.Text(), nullable=True),
-            sa.Column("processed_text", sa.Text(), nullable=True),
-            sa.Column(
-                "message_type",
-                sa.String(50),
-                nullable=False,
-                server_default=sa.text("'message'"),
-            ),
-            sa.Column("subtype", sa.String(50), nullable=True),
-            sa.Column(
-                "is_edited",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column("edited_ts", sa.String(50), nullable=True),
-            sa.Column(
-                "has_attachments",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "attachments", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("files", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-            sa.Column("thread_ts", sa.String(50), nullable=True),
-            sa.Column(
-                "is_thread_parent",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "is_thread_reply",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column(
-                "reply_count", sa.Integer(), nullable=False, server_default=sa.text("0")
-            ),
-            sa.Column(
-                "reply_users_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column(
-                "reaction_count",
-                sa.Integer(),
-                nullable=False,
-                server_default=sa.text("0"),
-            ),
-            sa.Column("message_datetime", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_analyzed",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("false"),
-            ),
-            sa.Column("message_category", sa.String(100), nullable=True),
-            sa.Column("sentiment_score", sa.Float(), nullable=True),
-            sa.Column(
-                "analysis_data", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-            ),
-            sa.Column("channel_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("user_id", postgresql.UUID(as_uuid=True), nullable=True),
-            sa.Column("parent_id", postgresql.UUID(as_uuid=True), nullable=True),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["channel_id"], ["slackchannel.id"]),
-            sa.ForeignKeyConstraint(["user_id"], ["slackuser.id"]),
-            sa.ForeignKeyConstraint(["parent_id"], ["slackmessage.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackmessage_id"), "slackmessage", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackmessage_slack_id"), "slackmessage", ["slack_id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackmessage_slack_ts"), "slackmessage", ["slack_ts"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackmessage_thread_ts"),
-            "slackmessage",
-            ["thread_ts"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackmessage_message_datetime"),
-            "slackmessage",
-            ["message_datetime"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackmessage_channel_id"),
-            "slackmessage",
-            ["channel_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackmessage_user_id"), "slackmessage", ["user_id"], unique=False
-        )
-        op.create_index(
-            "ix_slackmessage_channel_id_slack_ts",
-            "slackmessage",
-            ["channel_id", "slack_ts"],
-            unique=False,
-        )
-        op.create_index(
-            "ix_slackmessage_user_id_slack_ts",
-            "slackmessage",
-            ["user_id", "slack_ts"],
-            unique=False,
-        )
-
-        # SlackReaction Table
-        op.create_table(
-            "slackreaction",
-            sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("emoji_name", sa.String(255), nullable=False),
-            sa.Column("emoji_code", sa.String(255), nullable=True),
-            sa.Column("reaction_ts", sa.String(50), nullable=True),
-            sa.Column("message_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("user_id", postgresql.UUID(as_uuid=True), nullable=False),
-            sa.Column("created_at", sa.DateTime(), nullable=False),
-            sa.Column("updated_at", sa.DateTime(), nullable=False),
-            sa.Column(
-                "is_active",
-                sa.Boolean(),
-                nullable=False,
-                server_default=sa.text("true"),
-            ),
-            sa.ForeignKeyConstraint(["message_id"], ["slackmessage.id"]),
-            sa.ForeignKeyConstraint(["user_id"], ["slackuser.id"]),
-            sa.PrimaryKeyConstraint("id"),
-        )
-
-        op.create_index(
-            op.f("ix_slackreaction_id"), "slackreaction", ["id"], unique=False
-        )
-        op.create_index(
-            op.f("ix_slackreaction_message_id"),
-            "slackreaction",
-            ["message_id"],
-            unique=False,
-        )
-        op.create_index(
-            op.f("ix_slackreaction_user_id"), "slackreaction", ["user_id"], unique=False
-        )
-        op.create_index(
-            "ix_slackreaction_message_id_user_id_emoji_name",
-            "slackreaction",
-            ["message_id", "user_id", "emoji_name"],
-            unique=True,
-        )
-
-    # Create database tables with idempotency checks
-    create_table_if_not_exists("team", create_team_table)
-    create_table_if_not_exists("teammember", create_teammember_table)
-    create_table_if_not_exists("integration", create_integration_table)
-    create_table_if_not_exists(
-        "integration_credential", create_integration_credential_table
-    )
-    create_table_if_not_exists("integration_share", create_integration_share_table)
-    create_table_if_not_exists("service_resource", create_service_resource_table)
-    create_table_if_not_exists("resource_access", create_resource_access_table)
-    create_table_if_not_exists("integration_event", create_integration_event_table)
-
-    # Try to create Slack tables but first check if SlackAnalysis exists
-    conn = op.get_bind()
-    slackanalysis_exists_query = """
-    SELECT 1 FROM information_schema.tables 
-    WHERE table_name = 'slackanalysis' AND table_schema = 'public'
-    """
-    result = conn.execute(sa.text(slackanalysis_exists_query))
-    slackanalysis_exists = result.scalar() is not None
-
-    if not slackanalysis_exists:
-        try:
-            create_slack_tables()
-        except Exception as e:
-            # Some slack tables might already exist, which would cause errors
-            # but we can ignore them as long as integration tables work
-            print(f"NOTE: Error creating some Slack tables (might be ok): {e}")
-
-
-def downgrade():
-    """
-    This is a one-way schema consolidation.
-    In case a downgrade is needed, use individual migrations.
-    """
-    pass
