--- backend/alembic/versions/fe6424d2cda2_add_team_models.py
@@ -1,127 +1,0 @@
-"""add_team_models
-
-Revision ID: fe6424d2cda2
-Revises: 48a0524f84c5
-Create Date: 2025-04-14 10:30:32.536665
-
-"""
-
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "fe6424d2cda2"
-down_revision = "48a0524f84c5"
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "team",
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column("slug", sa.String(length=255), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("avatar_url", sa.String(length=1024), nullable=True),
-        sa.Column("team_size", sa.Integer(), nullable=False),
-        sa.Column("is_personal", sa.Boolean(), nullable=False),
-        sa.Column(
-            "team_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True
-        ),
-        sa.Column("created_by_user_id", sa.String(length=255), nullable=False),
-        sa.Column("created_by_email", sa.String(length=255), nullable=True),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("updated_at", sa.DateTime(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_team_id"), "team", ["id"], unique=False)
-    op.create_index(op.f("ix_team_slug"), "team", ["slug"], unique=False)
-    op.create_index("ix_team_slug_unique", "team", ["slug"], unique=True)
-    op.create_table(
-        "teammember",
-        sa.Column("user_id", sa.String(length=255), nullable=False),
-        sa.Column("email", sa.String(length=255), nullable=True),
-        sa.Column("display_name", sa.String(length=255), nullable=True),
-        sa.Column(
-            "role",
-            sa.Enum("OWNER", "ADMIN", "MEMBER", "VIEWER", name="teammemberrole"),
-            nullable=False,
-        ),
-        sa.Column("invitation_status", sa.String(length=50), nullable=False),
-        sa.Column("invitation_token", sa.String(length=255), nullable=True),
-        sa.Column("invitation_expires_at", sa.DateTime(), nullable=True),
-        sa.Column("last_active_at", sa.DateTime(), nullable=True),
-        sa.Column("team_id", sa.UUID(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("updated_at", sa.DateTime(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["team_id"],
-            ["team.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_teammember_id"), "teammember", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_teammember_team_id"), "teammember", ["team_id"], unique=False
-    )
-    op.create_index(
-        "ix_teammember_team_id_user_id",
-        "teammember",
-        ["team_id", "user_id"],
-        unique=True,
-    )
-    op.create_index(
-        op.f("ix_teammember_user_id"), "teammember", ["user_id"], unique=False
-    )
-    op.create_index(op.f("ix_slackanalysis_id"), "slackanalysis", ["id"], unique=False)
-    op.create_index(op.f("ix_slackchannel_id"), "slackchannel", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_slackchannelanalysis_id"), "slackchannelanalysis", ["id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_slackcontribution_id"), "slackcontribution", ["id"], unique=False
-    )
-    op.create_index(op.f("ix_slackmessage_id"), "slackmessage", ["id"], unique=False)
-    op.create_index(op.f("ix_slackreaction_id"), "slackreaction", ["id"], unique=False)
-    op.create_index(op.f("ix_slackuser_id"), "slackuser", ["id"], unique=False)
-    op.add_column("slackworkspace", sa.Column("team_id", sa.UUID(), nullable=True))
-    op.create_index(
-        op.f("ix_slackworkspace_id"), "slackworkspace", ["id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_slackworkspace_team_id"), "slackworkspace", ["team_id"], unique=False
-    )
-    op.create_foreign_key(None, "slackworkspace", "team", ["team_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "slackworkspace", type_="foreignkey")
-    op.drop_index(op.f("ix_slackworkspace_team_id"), table_name="slackworkspace")
-    op.drop_index(op.f("ix_slackworkspace_id"), table_name="slackworkspace")
-    op.drop_column("slackworkspace", "team_id")
-    op.drop_index(op.f("ix_slackuser_id"), table_name="slackuser")
-    op.drop_index(op.f("ix_slackreaction_id"), table_name="slackreaction")
-    op.drop_index(op.f("ix_slackmessage_id"), table_name="slackmessage")
-    op.drop_index(op.f("ix_slackcontribution_id"), table_name="slackcontribution")
-    op.drop_index(op.f("ix_slackchannelanalysis_id"), table_name="slackchannelanalysis")
-    op.drop_index(op.f("ix_slackchannel_id"), table_name="slackchannel")
-    op.drop_index(op.f("ix_slackanalysis_id"), table_name="slackanalysis")
-    op.drop_index(op.f("ix_teammember_user_id"), table_name="teammember")
-    op.drop_index("ix_teammember_team_id_user_id", table_name="teammember")
-    op.drop_index(op.f("ix_teammember_team_id"), table_name="teammember")
-    op.drop_index(op.f("ix_teammember_id"), table_name="teammember")
-    op.drop_table("teammember")
-    op.drop_index("ix_team_slug_unique", table_name="team")
-    op.drop_index(op.f("ix_team_slug"), table_name="team")
-    op.drop_index(op.f("ix_team_id"), table_name="team")
-    op.drop_table("team")
-    # ### end Alembic commands ###
